selectJustsM -> selectJusts
Made selectJustsM into a pure function.
selectJustsM wasn't doing any work in the IO monad, so we shouldn't
confuse our future selves and introduce room for error by forcing it to.

Pure functions can be lifted to operate on monadic values with fmap or
liftM. Consider:

fmap :: Functor f
     => (a -> b)
     -> (f a -> f b)

The above is one possible way to read the type signature for fmap, it
puts emphasis on the interpretation that fmap takes a function that
operates on pure values and lifts it to work in a certain context f.

Commit 1 of 2 on selectJustsM.