Replaced selectJusts with catMaybes
catMaybes from Data.Maybe is a functionally equivalent version of
selectJusts. This is one further advantage of "shrinking" the types and
scopes of functions to just what is necessary, it becomes easy to search
Hoogle for function signatures and find where functionality has already
been implemented by standard library functions.

http://www.haskell.org/hoogle/?hoogle=[Maybe+a]+-%3E+[a]

catMaybes is the first result for searching for the pure version of the
selectJusts type signature.

This is commit 2 of 2 on selectJusts.