Reimplemented randomElem
randomElem was much more complicated than it needs to be, after getting
the random index for an element, you can use !! for list indexing
instead of adding indices to the entire list and filtering.

This is another standard library function that you could search for in
Hoogle. Take a plain English explanation of what you need a function to
do, convert it to a type signature, then search for it. E.g.,

Given a list and an index, return the element at that index.
becomes
[a] -> Int -> a