Replaced shuffle with shuffleM from random-shuffle
There isn't a list shuffling function provided in anywhere in the base
package, but since it's one of those things that's easier to get subtly
wrong than obviously right I added a dependency on the random-shuffle
package and replaced shuffle with shuffleM (which operates in an
arbitrary MonadRandom monad).

The original implementation of shuffle had O(n^2) complexity, and
was weirdly biased due to the Eq constraint. Consider this:

data Foo = Foo Bool Int
instance Eq Foo where
    (Foo p _) == (Foo q _) = p == q

shuffle [Foo True 1, Foo True 2]
--> suppose randomElem selects Foo True 2
--> re = Foo True 2; restList = [Foo True 2]
== [Foo True 2, Foo True 2]

That may be a pathologic instance of Eq, but even so you wouldn't expect
a shuffle function to drop or duplicate elements.

After switching shuffle functions we can also delete the randomElem
function.