Revises randomFillShifts
First this reorders clauses to match the order of fields within the
Shift datatype (i.e., shift designator first, person second). That
involves changing the where clause to a let and moving it within the
body of the do block.

Next the use of zip and map in pairing up people with shifts was changed
to use zipWith. zipWith is like a generalized version of zip, you just
provide it with a joining function on two elements and it lifts the
function to work on lists of those elements. In this case we can lift
the Shift constructor to work on lists of shifts and Persons.