Revises randomDay
The point of this version of randomDay is to highlight the fact that
this function is really a wrapper around randomFillShifts. With some
more thought about control flow this function could probably be
eliminated.

Notice how I once again use fmap to lift the Day constructor (with its
first argument applied) into the IO context to wrap the value returned
from the call of randomFillShifts. Without the do-syntax around,
hopefully this looks as thin as it is.