Revises getStaffList
First, makes the function total by accounting for the case of the empty
Schedule, []. It's a good idea to use and write as many total functions
as possible in Haskell, this reduces your surface area for errors.

Also replaces the use of an anonymous lambda with a locally defined
helper function. This is a bit more readable.