Revises longestStreak
Rewritten using a pipeline of standard list functions.

This version nicely separates concerns into separate stages. First it
groups elements by equality, then filters out groups which do not match
the given element, then finds the run length of each remaining group,
prepends the list of run lengths with the identity element (0) so that
it will never be empty and cause `maximum` to emit an error, and finally
takes the maximum value from the list.